Dashboard refactor/450 move rh toolbar to component (#451)
* moved basics to own component - no func as of now

* added outside visibilit

* added functionality - needs proper testing

* fixed mission items filtering

* all functionality works

* formatted

* changed wording
