+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "cSpell.words": [
         "ARSP",
         "centered",
+        "Crosshair",
         "falcongrey",
         "falconred",
         "maplibre",
